Revert "Breaking out resource table + functionality to its own component." Re-Applying new resource removal tracking method. (72d6ac296e94b96f2c2c825615bd44364d9cfebe)
This reverts commit 2ad7923befafed980a1eeddf2a15462e1207c75e.
